refactor(server): break the 9 noImportCycles cycles, pin the rule to error (#1005)
Extract `surfaceCtx.ts` as a Proxy-fronted holder for the typed
mutation context. `surface.ts` registers via `setSurfaceCtx(built)`
once at startup; `activity.ts`, `session.ts`, `terminalBackend/local.ts`,
and `terminalBackend/metadata.ts` import `surfaceCtx` from the holder.
The bidirectional `surface.ts ↔ domain` arrow collapses to a one-way
arrow plus a one-way registration — same shape as PR #998's cycle 1.

Move `unwrapGit` to its own `unwrapGit.ts` so `terminalBackend/local.ts`
no longer needs any import from `surface.ts`; `router.ts` follows.

With zero cycles, promote `lint/suspicious/noImportCycles` from biome's
default `warn` to `error` in `biome.jsonc`. The TDZ-on-import-reorder
class that bit #1003 in `just smoke` (vite-node masked the bug; only
the production Node ESM loader caught it) is now structurally
impossible.

The load-order `biome-ignore-start assist/source/organizeImports`
markers in `surface.ts` also retire — biome's alphabetical sort no
longer matters once there's no cycle for the order to converge along.
