+++ packages/server/src/router.ts
@@ -23,10 +23,11 @@ import { match } from "ts-pattern";
 import { serverHostname, serverProcessId } from "./hostname.ts";
 import { log } from "./log.ts";
 import { pwaIdentityForHostname } from "./pwaIdentity.ts";
-import { surfaceRouter, t, unwrapGit } from "./surface.ts";
+import { surfaceRouter, t } from "./surface.ts";
 import { getTerminal, type TerminalProcess } from "./terminal-registry.ts";
 import { getTerminalBackendFor } from "./terminalBackend/index.ts";
 import { saveTerminalFile } from "./terminalScratch.ts";
+import { unwrapGit } from "./unwrapGit.ts";
 import {
   createTerminal,
   killAllTerminals,
